Extract and generify oracle key conversion
